requirements.txt
APScheduler==3.10.4
beautifulsoup4==4.12.3
feedparser==6.0.11
fastapi==0.110.0
google-api-python-client==2.121.0
google-auth==2.28.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
jinja2==3.1.3
pydantic==2.6.3
python-dotenv==1.0.1
requests==2.31.0
sqlalchemy==2.0.27
tenacity==8.2.3
uvicorn==0.27.1
